+++ src/seamlesserverpkg/renderer/base/bridgedData.nim
@@ -1,3 +1,5 @@
+from std/times import Duration, DateTime
+
 include pkg/karax/prelude
 from pkg/prologue/core/types import FlashLevel
 
@@ -8,7 +10,11 @@ type
   FlashMessage* = object
     level*: FlashLevel
     text*: string
+    created*: int64 ## unix time
+    duration*: int ## ms
   BridgedData* = ref object
+    ## The data sent by backend to frontend;  
+    ## Needs to be able to serialize and deserialize in JSON
     appName*: string
     flashes*: seq[FlashMessage]
     devData*: string
@@ -28,9 +34,6 @@ proc newBridgedData*: BridgedData =
 when not defined js:
   from std/json import `$`, `%*`, `%`, JsonNode
 
-  # proc `%`*(e: FlashLevel): JsonNode =
-  #   result = JsonNode(kind: JString, str: $e)
-  
   func genBridgedData*(self): VNode =
     ## Generates the data that'll be sent to Javascript backend
     let node = $ %*self