+++ src/seamlesserverpkg/userMessages.nim
@@ -1,0 +1,4 @@
+const
+  umInvalidEmail* = "Invalid email"
+  umInvalidPassword* = "Invalid password. Make sure that's 6 characters up to 32."
+  umInvalidUsername* = "Invalid username. Make sure that's smaller than 16 characters."