+++ src/seamlesserverpkg/validation/username.nim
@@ -1,0 +1,7 @@
+const pattern* = r"""^.{,16}$"""
+
+when not defined js:
+  import std/re
+  proc check*(username: string): bool =
+    ## Checks if username is correct
+    username.match pattern.re